+++ package.json
@@ -14,6 +14,7 @@
     "axios": "^1.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.16.0",
     "typeface-inter": "^3.18.1"
   },
   "devDependencies": {
