+++ src/pages/Home.jsx
@@ -1,0 +1,19 @@
+import HeroSection from "../components/HeroSection/HeroSection";
+import FAQSection from "../components/FAQSection/FAQSection";
+// import Sponsors from "./pages/SponsorSection/Sponsors.jsx";
+import Explore from "../components/exploreSection/explore";
+import RegisterSection from "../components/RegisterSection/registerSection";
+
+function Home() {
+  // do not add css in this file
+  return (
+    <div>
+      <HeroSection />
+      <Explore />
+      <FAQSection />
+      <RegisterSection />
+    </div>
+  );
+}
+
+export default Home;
