--- src/pages/SponsorSection/Sponsors.jsx
@@ -1,18 +1,0 @@
-import CommunitySection from "./CommunitySection.jsx";
-import CollabSection from "./CollabSection.jsx";
-import SpecialSection from "./SpecialSection.jsx";
-import SponsoredBy from "./SponsoredBy.jsx";
-import CloudSection from "./CloudSection.jsx";
-function Sponsors() {
-  return (
-    <section className="">
-      <CommunitySection />
-      <CollabSection  />
-      <SpecialSection />
-      <SponsoredBy />
-      <CloudSection />
-    </section>
-  );
-}
-
-export default Sponsors;
