+++ src/pages/Sponsors.jsx
@@ -1,0 +1,18 @@
+import CommunitySection from "../components/SponsorSection/CommunitySection.jsx";
+import CollabSection from "../components/SponsorSection/CollabSection.jsx";
+import SpecialSection from "../components/SponsorSection/SpecialSection.jsx";
+import SponsoredBy from "../components/SponsorSection/SponsoredBy.jsx";
+import CloudSection from "../components/SponsorSection/CloudSection.jsx";
+function Sponsors() {
+  return (
+    <section className="">
+      <CommunitySection />
+      <CollabSection />
+      <SpecialSection />
+      <SponsoredBy />
+      <CloudSection />
+    </section>
+  );
+}
+
+export default Sponsors;
